Implemented csv export; other bugfixes

+++ src/org/hiphap/Screens/ManageUsers.java
@@ -5,9 +5,10 @@ import org.hiphap.UserManager;
 
 public class ManageUsers extends MenuScreen {
   public ManageUsers() {
-    addMenuOption("1", "Change current user's password");
-    addMenuOption("2", "Create new user");
-    addMenuOption("3", "Delete user");
+    addMenuOption("1", "List system users");
+    addMenuOption("2", "Change current user's password");
+    addMenuOption("3", "Create new user");
+    addMenuOption("4", "Delete user");
   }
 
   @Override
@@ -20,13 +21,20 @@ public class ManageUsers extends MenuScreen {
   Transition handleInput(String input) {
     switch (input) {
       case "1":
-        return changePassword();
+        clearScreen();
+        for (String username: UserManager.getInstance().getUsernames()) {
+          System.out.printf(" - %s%n", username);
+        }
+        readString("Any input to continue...\n");
+        return new Transition(Transition.Type.SUCCESS);
       case "2":
-        return createNewUser();
+        return changePassword();
       case "3":
+        return createNewUser();
+      case "4":
         return deleteUser();
       default:
-        return new Transition(Transition.Type.INVALID, "Invalid input; try again.");
+        return new Transition(Transition.Type.ERROR, "Invalid input; try again.");
     }
   }
 
@@ -37,7 +45,7 @@ public class ManageUsers extends MenuScreen {
     if (UserManager.getInstance().getCurrentUser().changePassword(oldPassword, newPassword, repeatPassword)) {
       return new Transition(Transition.Type.SUCCESS, "Password successfully changed.");
     } else {
-      return new Transition(Transition.Type.INVALID, "Wrong password, or new passwords do not match.");
+      return new Transition(Transition.Type.ERROR, "Wrong password, or new passwords do not match.");
     }
   }
 
@@ -57,7 +65,7 @@ public class ManageUsers extends MenuScreen {
       UserManager.getInstance().addUser(new User(username, password, isAdmin));
       return new Transition(Transition.Type.SUCCESS, "User successfully added.");
     } else {
-      return new Transition(Transition.Type.INVALID, "Username already exists.");
+      return new Transition(Transition.Type.ERROR, "Username already exists.");
     }
   }
 
@@ -65,16 +73,20 @@ public class ManageUsers extends MenuScreen {
     String username = clsAndReadString("Which user would you like to delete? ");
     if (UserManager.getInstance().isUsernameTaken(username)) {
       if (!UserManager.getInstance().getCurrentUser().getUsername().equals(username)) {
+        if (!readBoolean("WARNING! This action cannot be undone. Are you sure " +
+            "you want to delete the user \"" + username + "\"?")) {
+          return new Transition(Transition.Type.ERROR, "Operation aborted.");
+        }
         if (UserManager.getInstance().deleteUser(username)) {
           return new Transition(Transition.Type.SUCCESS, "User deleted successfully.");
         } else {
-          return new Transition(Transition.Type.INVALID, "Error deleting user.");
+          return new Transition(Transition.Type.ERROR, "Error deleting user.");
         }
       } else {
-        return new Transition(Transition.Type.INVALID, "Cannot delete the currently logged in user.");
+        return new Transition(Transition.Type.ERROR, "Cannot delete the currently logged in user.");
       }
     } else {
-      return new Transition(Transition.Type.INVALID, "Username does not exist.");
+      return new Transition(Transition.Type.ERROR, "Username does not exist.");
     }
   }
 }
